requirements.txt
asgi_redis
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==7.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1
django-channels-graphql-ws==0.9.1
django-cors-headers==3.13.0
django-filter==22.1
django-graphql-jwt==0.3.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
frozenlist==1.3.1
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-ws==0.4.4
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
multidict==6.0.2
Pillow==9.2.0
promise==2.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-decouple==3.6
pytz==2022.2.1
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
text-unidecode==1.3
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
yarl==1.8.1
zope.interface==5.4.0
